Run make cli as current user
Any files generated by commands the user would run in the make cli would
be owned as root. This causes editors to be unable to modify those files
without elevated privileges.

This commit changes make cli to add the current user to the docker
container and start up a bash process as that user so that modifications
to files inside the container continue to be editable outside the
container.